Move TypeOfAny to mypyc_hacks
